[mod](Framework): convert parameter type in draw function

+++ common/include/d3d12mesh.hpp
@@ -190,7 +190,7 @@ public:
      */
     void draw(ID3D12GraphicsCommandList* pCmdList) const;
 
-    void draw(ID3D12GraphicsCommandList* pCmdList, std::uint32_t instanceCount) const;
+    void draw(ID3D12GraphicsCommandList* pCmdList, std::size_t instanceCount) const;
 
 private:
     void buildRes(d3d12::Core& core, d3d12::D3D12RenderContext& ctx, const VertexBuffer& vb, const IndexCont& ib);
